+++ changelog.md
@@ -5,10 +5,11 @@
 - `weui` 版本更新到 `0.4.2`, `weui` 官方的更新日志请参考 [https://github.com/weui/weui/blob/master/CHANGELOG.md](https://github.com/weui/weui/blob/master/CHANGELOG.md).
 - 大幅增强 `Select` 组件:
   - 增加了 `update, open, close` 方法，现在可以在初始化之后通过 `$(xx).select("update", config)` 动态修改任何配置。
-  - 增加了 `onOpen, onClose` 回调
-- 增加了新组件 `Photo Browser`,暂时还不支持手势缩放
+  - 增加了 `onOpen, onClose` 回调.
+- 增加了新组件 `Photo Browser`,暂时还不支持手势缩放.
 - `Action Sheet` 增加了 `color-*, bg-*` 可以配置不同的颜色和背景色，增加了 `onClose` 回调函数。
-- 修复了 `datetime picker` 在 ios 上设置最大最小值无效的问题
+- 修复了 `datetime picker` 在 ios 上设置最大最小值无效的问题。
+- 修复了 `popup` 会被 `select` 和 `navbar` 挡住的问题。
 
 # V0.6.1 - 2016/04/20